complete shooting for player
