+++ Assets/Scripts/Managers/BulletManager.cs
@@ -7,7 +7,7 @@ using UnityEngine.Pool;
 public class BulletManager : MonoBehaviour
 {
     [SerializeField] Bullet bullletPref;
-    private static IObjectPool<Bullet> bulletPool;
+    public IObjectPool<Bullet> bulletPool;
     [SerializeField] private bool collectionCheck;
 
     //Singleton
@@ -34,12 +34,13 @@ public class BulletManager : MonoBehaviour
 
     private void ReleaseBullet(Bullet bullet)
     {
-        bullet.enabled = false;
+        Debug.Log("bullet released");
+        bullet.gameObject.SetActive(false);
     }
 
     private void GetBullet(Bullet bullet)
     {
-        bullet.enabled = true;
+        bullet.gameObject.SetActive(true);
     }
     private void DestroyBullet(Bullet bullet)
     {
