+++ Assets/Scripts/Units/Health/PlayerHealth.cs
@@ -4,8 +4,6 @@ using UnityEngine;
 
 public class PlayerHealth : BaseHealth
 {
-    public float offset = 1f; // shoot offset
-
     private void Awake()
     {
         this.gameObject.tag = "Allie";
@@ -13,7 +11,8 @@ public class PlayerHealth : BaseHealth
 
     public void Absorb(Bullet bullet)
     {
-        bullet.gameObject.SetActive(false);
+        BulletManager.Instance.bulletPool.Release(bullet);
+        //bullet.gameObject.SetActive(false);
         Heal(1);
     }
 
@@ -22,23 +21,6 @@ public class PlayerHealth : BaseHealth
         if (tag == this.gameObject.tag)
         {
             Absorb(bullet);
-        }
-    }
-    private void Update()
-    {
-        //pseudo
-        if (Input.GetKeyDown(KeyCode.Mouse0))
-        {
-            Bullet bullet = BulletManager.Instance.GetBullet(this.gameObject.tag);
-            Vector2 direction = (Vector2)Camera.main.ScreenToWorldPoint(Input.mousePosition) - (Vector2)transform.position;
-            // Apply the offset along the direction
-            Vector2 offsetPosition = (Vector2)this.transform.position + direction.normalized * offset;
-
-            // Set the bullet's position
-            bullet.transform.position = offsetPosition;
-            bullet.Shoot(direction);
-
-            Debug.Log("Shooting");
         }
     }
 }
