+++ Assets/Scripts/Units/ShootProjectiles.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fa1abdb1a49af6d4497a5f92065bc31c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
